converters/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
beautifulsoup4==4.12.3
camelot-py==0.9.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
EasyProcess==1.1
entrypoint2==1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
fastapi==0.110.2
filelock==3.14.0
gdown==5.1.0
ghostscript==0.7
h11==0.14.0
idna==3.7
lxml==5.2.1
numpy==1.26.4
opencv-python==4.9.0.80
openpyxl==3.1.2
orjson==3.10.1
pandas==2.2.2
patool==2.2.0
pdfminer.six==20231228
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pypdf==4.2.0
PyPDF2==2.12.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-multipart==0.0.9
pytz==2024.1
pyunpack==0.3
requests==2.31.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
xlrd==2.0.1
